Reuse existing CopyTextureToTexture for depth texture copy
