--- packages/dev/core/src/Shaders/fluidRenderingPassDepth.fragment.fx
@@ -1,8 +1,0 @@
-uniform sampler2D textureDepth;
-
-varying vec2 vUV;
-
-void main(void) 
-{
-	gl_FragDepth = texture2D(textureDepth, vUV).x;
-}
